# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Site API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.local",
            "description": "Local HTTPS server"
        }
    ],
    "paths": {
...
        "/api/vk/send": {
            "get": {
                "tags": [
                    "VK"
                ],
                "summary": "Отправить сообщение пользователю VK (через GET)",
                "description": "То же, что и POST, но параметры передаются в query-строке.",
                "operationId": "apiVkSendMessageGet",
                "parameters": [
                    {
                        "name": "group_name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "main_group"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "vk1.a.xxxxx"
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "12345,67890"
                    },
                    {
                        "name": "message",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Привет из API"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение отправлено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "group": {
                                            "type": "string",
                                            "example": "main_group",
                                            "nullable": true
                                        },
                                        "sent_user_ids": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "VK"
                ],
                "summary": "Отправить сообщение пользователю VK от имени группы (через POST)",
                "description": "Для авторизации передайте либо название группы, либо token. Подробности в описании.",
                "operationId": "apiVkSendMessagePost",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message"
                                ],
                                "properties": {
                                    "group_name": {
                                        "type": "string",
                                        "example": "main_group",
                                        "nullable": true
                                    },
                                    "token": {
                                        "type": "string",
                                        "example": "vk1.a.xxxxx",
                                        "nullable": true
                                    },
                                    "user_id": {
                                        "description": "ID пользователя: число, массив ID или строка \"1,2,3\"",
                                        "nullable": true,
                                        "oneOf": [
                                            {
                                                "type": "integer",
                                                "example": 12345
                                            },
                                            {
                                                "type": "string",
                                                "example": "12345,67890"
                                            },
                                            {
                                                "type": "array",
                                                "items": {
                                                    "type": "integer"
                                                }
                                            }
                                        ]
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "Привет из API"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Сообщение отправлено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "group": {
                                            "type": "string",
                                            "example": "main_group",
                                            "nullable": true
                                        },
                                        "sent_user_ids": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации или доменная ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Token for group not found: main_group"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}